requirements.txt
Flask==2.0.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
gunicorn==21.2.0
requests==2.31.0
tqdm==4.66.2
gdown==4.7.3
